Attempt at splitting into different classes using mixins
Not sure this is the right approach, it's quite complex.

The hard part is we begin with the device being a `Device`, but then once we call some methods on it we get more info about product ID, version, label etc.

So instead I did device.fixture = Class (eg Light, Switch), but then I want to have a class per set of functionalities, so I used mixins (e.g. Hev, Multizone, Color).

But it becomes hard to use this in __main__.py. There should be a way to narrow the type so only methods applicable to that given fixture are available. I can't figure out the right data structure thoguh